% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip.R
\name{calculate_Fws}
\alias{calculate_Fws}
\title{Calculate the complexity of infection in every sample (Fws)}
\usage{
calculate_Fws(snpdata)
}
\arguments{
\item{snpdata}{a SNPdata object}
}
\value{
a SNPdata object with an updated meta table
}
\description{
Fws is the within host genetic diversity
}
