% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip.R
\name{compute_MAF}
\alias{compute_MAF}
\title{Calculate minor allele frequency (MAF)}
\usage{
compute_MAF(snpdata, include.het = TRUE, mat.name = "GT")
}
\arguments{
\item{snpdata}{a SNPdata object}

\item{include.het}{whether to account for the heterozygous allele or not. this is only used when mat.name="GT"}

\item{mat.name}{the name of the matrix to use. default is "GT"}
}
\value{
a SNPdata object with an updated details table
}
\description{
Uses the SNPdata object to calculate the MAF at every loci
}
\details{
if include.het=FALSE, the mixed allele will not be considered in the MAF calculation
}
