% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip.R
\name{drop_snps}
\alias{drop_snps}
\title{Drop set a SNPs}
\usage{
drop_snps(snpdata, snp.to.be.dropped)
}
\arguments{
\item{snpdata}{a SNPdata object}

\item{snp.to.be.dropped}{a data frame with 2 columns "Chrom" and "Pos"}
}
\value{
a SNPdata object where the specified SNPs have been removed
}
\description{
remove a set of SNPs from the SNPdata object
}
