% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip.R
\name{select_chrom}
\alias{select_chrom}
\title{Select data from specified chromosomes}
\usage{
select_chrom(snpdata, chrom = "all")
}
\arguments{
\item{snpdata}{a SNPdata object}

\item{chrom}{a comma-separated list of chromosomes}
}
\value{
a SNPdata object with only the data from the specified chromosomes
}
\description{
return data for specified chromosomes only
}
