% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip.R
\name{snp_get_tables}
\alias{snp_get_tables}
\title{Generate the SNPdata object}
\usage{
snp_get_tables(
  vcf.file = RAW,
  meta.file = METADATA,
  output.dir = PROCESSED,
  min.mq = 30,
  min.dp = 4
)
}
\arguments{
\item{vcf.file}{the input VCF file}

\item{meta.file}{the metadata file}

\item{output.dir}{the path to the folder where to store the output files}

\item{min.mq}{the minimum MAPQ (read mapping quality)}

\item{min.dp}{the minimum DP (read depth)}
}
\value{
an object of class SNPdata.
}
\description{
This function generate the input data needed for whole genome SNP data genotyped from malaria parasite
}
\details{
use the print.SNPdata() function to print the created object
}
