% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip.R
\name{get_snpdata}
\alias{get_snpdata}
\title{Generate the SNPdata object}
\usage{
get_snpdata(vcf.file = NA, meta.file = NA, output.dir = NA)
}
\arguments{
\item{vcf.file}{the input VCF file}

\item{meta.file}{the metadata file}

\item{output.dir}{the path to the folder where to store the output files}
}
\value{
an object of class SNPdata.
}
\description{
This function generate the input data needed for whole genome SNP data genotyped from malaria parasite
}
\details{
use the print.SNPdata() function to print the created object
}
